requirements.txt
attrs==21.2.0
beancount==2.3.4
beautifulsoup4==4.9.3
bottle==0.12.19
cachetools==4.2.2
cchardet==2.1.7
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
dateparser==1.0.0
eml-parser==1.14.7
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
iniconfig==1.1.1
lxml==4.6.3
packaging==21.0
pluggy==0.13.1
ply==3.11
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.10.1
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-magic==0.4.24
pytz==2021.1
pyzipper==0.3.5
regex==2021.7.6
requests==2.26.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
xlrd==2.0.1
camelot-py==0.11.0
opencv-python==4.8.0.74
